chore(pack): pull out parser
