+++ pathing/pack/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "taimi_pack"
+authors = [ "kittywitch" ]
+description = "GW2 TacO packs"
+version = "0.1.0"
+edition = "2021"
+
+[features]
+default = [
+  "zip",
+]
+zip = [
+  "dep:zip",
+]
+
+[lints.rust.unexpected_cfgs]
+level = "warn"
+check-cfg = [
+  'cfg(todo, values("unused"))',
+  'cfg(todo)',
+]
+
+[dependencies]
+anyhow = "1"
+log = "0.4"
+# TODO? tokio = { version = "1", features = [ "fs", "macros" ] }
+glam = "0.30"
+glamour = "0.18"
+indexmap = "2.9"
+relative-path = "1"
+xml-rs = "0.8"
+uuid = { version = "1.17", features = ["v4"] }
+md5 = "0.7"
+base64 = "0.22"
+zip = { version = "4", optional = true }
+bitvec = "1.0.1"
