+++ pathing/pack/src/lib.rs
@@ -1,0 +1,18 @@
+pub mod attributes;
+pub mod loader;
+pub mod category;
+pub mod pack;
+pub mod poi;
+pub mod trail;
+
+pub use {
+    self::{
+        attributes::MarkerAttributes,
+        pack::Pack,
+        category::Category,
+        loader::{PackLoaderContext, LoaderAssetReader},
+        poi::Poi,
+        trail::Trail,
+    },
+    uuid::Uuid,
+};
