+++ src/space/pack/mod.rs
@@ -1,317 +1,89 @@
 use {
-    super::{
-        dx11::{PerspectiveHandler, PerspectiveInputData, RenderBackend},
-        render_list::{MapFrustum, RenderEntity, RenderId, RenderList, RenderListBuilder},
-        resources::Texture,
-    },
-    crate::marker::atomic::MapSpace,
+    croner::Cron,
     anyhow::Context,
-    bitvec::vec::BitVec,
-    category::Category,
-    glam::Vec3,
-    glamour::{Point3, Vector3},
-    indexmap::{map::Entry, IndexMap, IndexSet},
-    loader::{DirectoryLoader, PackLoaderContext, ZipLoader},
-    poi::{ActivePoi, PoiCommonRenderData},
-    std::{
-        collections::{HashMap, HashSet},
-        fs::{create_dir_all, read_dir},
-        io::{Cursor, Read as _},
-        path::Path,
-        sync::Arc,
-    },
-    trail::ActiveTrail,
-    uuid::Uuid,
-    windows::Win32::Graphics::Direct3D11::{ID3D11Device, ID3D11DeviceContext},
-    xml::{common::Position, reader::XmlEvent},
+    glamour::{Box3, Point3, Vector3},
 };
 
-pub mod activemod;
-
-pub mod attributes;
-pub mod category;
-pub mod loader;
 pub mod poi;
 pub mod trail;
 
+mod pack;
+
 pub use {
-    self::activemod::{
-        PackCollection,
-        ActivePack
+    taimi_pack::{
+        attributes::{self, MarkerAttributes},
+        category::{self, Category},
+        loader,
+        poi::Poi,
+        trail::{Trail, TrailSection},
+        pack::Pack,
+    },
+    self::{
+        pack::{
+            PackCollection,
+            ActivePack
+        },
+        poi::ActivePoi,
+        trail::ActiveTrail,
     },
 };
 
-#[derive(Default)]
-pub struct Pack {
-    pub name: String,
+pub type PackSpace = crate::marker::atomic::LocalSpace;
 
-    // Descriptive data.
-    pub pois: Vec<poi::Poi>,
-    pub trails: Vec<trail::Trail>,
-    pub categories: CategoryCollection,
+pub trait MarkerAttributesExt {
+    fn parse_schedule(&self) -> anyhow::Result<Option<Cron>>;
 }
 
-impl Pack {
-    pub fn load<L: PackLoaderContext>(loader: &mut L) -> anyhow::Result<Pack> {
-        let mut pack = Pack::default();
-
-        let pack_defs = loader.all_files_with_ext("xml")?;
-        for def in pack_defs {
-            parse_pack_def(&mut pack, loader, &def)?;
-        }
-
-        merge_category_attributes(&mut pack);
-        apply_marker_attributes(&mut pack);
-
-        Ok(pack)
-    }
-}
-
-#[derive(Default)]
-pub struct CategoryCollection {
-    /// Map full_id -> Category
-    pub all_categories: IndexMap<String, Category>,
-    /// List of root categories.
-    pub root_categories: IndexSet<String>,
-}
-
-fn taco_safe_name(value: &str, is_full: bool) -> String {
-    let mut result = String::with_capacity(value.len());
-    for c in value.chars() {
-        if c.is_ascii_alphanumeric() || (is_full && c == '.') {
-            result.push(c);
-        } else {
-            result.push('_');
-        }
-    }
-    result
-}
-
-/// I hate this. See: https://github.com/blish-hud/Pathing/blob/main/Utility/AttributeParsingUtil.cs#L39
-fn taco_xml_to_guid(value: &str) -> Uuid {
-    use base64::{engine::general_purpose, Engine as _};
-    let mut raw_guid = [0u8; 16];
-    if let Ok(len) = general_purpose::STANDARD.decode_slice(value, &mut raw_guid) {
-        if len == 16 {
-            return Uuid::from_bytes_le(raw_guid);
-        }
-    }
-    Uuid::from_bytes_le(md5::compute(value).0)
-}
-
-pub fn parse_pack_def(
-    pack: &mut Pack,
-    ctx: &mut impl PackLoaderContext,
-    asset: &str,
-) -> anyhow::Result<()> {
-    let mut stream = ctx.load_asset(asset)?;
-    let mut buf = Vec::new();
-    stream.read_to_end(&mut buf)?;
-    let data = String::from_utf8_lossy(&buf);
-    let mut parser = xml::EventReader::new(Cursor::new(data.into_owned().into_bytes()));
-
-    match inner_parse_pack_def(pack, ctx, &mut parser) {
-        Ok(()) => Ok(()),
-        Err(e) => Err(e).context(format!("Parsing pack def at {asset}:{}", parser.position())),
-    }
-}
-
-fn merge_category_attributes(pack: &mut Pack) {
-    for id in &pack.categories.root_categories {
-        inner_merge_category_attributes(&mut pack.categories.all_categories, id);
-    }
-}
-
-fn inner_merge_category_attributes(categories: &mut IndexMap<String, Category>, parent: &str) {
-    let attrs = categories[parent].marker_attributes.clone();
-    let children = categories[parent].sub_categories.clone();
-    for (_, id) in &*children {
-        if let Some(category) = categories.get_mut(id) {
-            Arc::make_mut(&mut category.marker_attributes).merge(&attrs);
-        } else {
-            log::error!("Inconsistent internal state, missing category `{id}`");
-            continue;
-        }
-        inner_merge_category_attributes(categories, id);
-    }
-}
-
-fn apply_marker_attributes(pack: &mut Pack) {
-    for poi in &mut pack.pois {
-        let Some(category) = pack.categories.all_categories.get(&poi.category) else {
-            continue;
-        };
-        poi.attributes.merge(&category.marker_attributes);
-    }
-    for trail in &mut pack.trails {
-        let Some(category) = pack.categories.all_categories.get(&trail.category) else {
-            continue;
-        };
-        trail.attributes.merge(&category.marker_attributes);
-    }
-}
-
-fn inner_parse_pack_def(
-    pack: &mut Pack,
-    ctx: &mut impl PackLoaderContext,
-    parser: &mut xml::EventReader<impl std::io::Read>,
-) -> anyhow::Result<()> {
-    let mut parse_stack: Vec<PartialItem> = Vec::with_capacity(16);
-
-    loop {
-        match parser.next()? {
-            XmlEvent::StartElement {
-                name, attributes, ..
-            } if valid_elem_start(parse_stack.last(), &name) => {
-                match name.local_name.to_ascii_lowercase().as_str() {
-                    "overlaydata" => {
-                        parse_stack.push(PartialItem::OverlayData);
-                    }
-                    "markercategory" => {
-                        let category = Category::from_xml(pack, &parse_stack, attributes)?;
-                        parse_stack.push(PartialItem::MarkerCategory(category));
-                    }
-                    "pois" => {
-                        parse_stack.push(PartialItem::PoiGroup);
-                    }
-                    "poi" => match poi::Poi::from_xml(pack, attributes) {
-                        Ok(poi) => parse_stack.push(PartialItem::Poi(poi)),
-                        Err(e) => {
-                            log::warn!("POI parse failed: {e:?}");
-                            parse_stack.push(PartialItem::PoisonElem);
-                        }
-                    },
-                    "trail" => match trail::Trail::from_xml(pack, ctx, attributes) {
-                        Ok(trail) => parse_stack.push(PartialItem::Trail(trail)),
-                        Err(e) => {
-                            log::warn!("Trail parse failed: {e:?}");
-                            parse_stack.push(PartialItem::PoisonElem);
-                        }
-                    },
-                    _ => anyhow::bail!("Unexpected <{name}>"),
-                }
-            }
-            XmlEvent::StartElement { name, .. } => anyhow::bail!("Unexpected <{name}>"),
-            XmlEvent::EndElement { .. }
-                if parse_stack.last().map(|i| i.is_poison()).unwrap_or(false) =>
-            {
-                parse_stack.pop();
-            }
-            XmlEvent::EndElement { name } if valid_elem_end(parse_stack.last(), &name) => {
-                match name.local_name.to_ascii_lowercase().as_str() {
-                    "overlaydata" | "pois" => {
-                        parse_stack.pop();
-                    }
-                    "markercategory" => {
-                        let Some(PartialItem::MarkerCategory(category)) = parse_stack.pop() else {
-                            anyhow::bail!("Inconsistent internal state");
-                        };
-
-                        match parse_stack.last_mut() {
-                            Some(PartialItem::OverlayData) => {
-                                pack.categories
-                                    .root_categories
-                                    .insert(category.full_id.clone());
-                            }
-                            Some(PartialItem::MarkerCategory(parent)) => {
-                                let subs = Arc::make_mut(&mut parent.sub_categories);
-                                subs.insert(category.id.clone(), category.full_id.clone());
-                            }
-                            _ => anyhow::bail!("Inconsistent internal state"),
-                        }
-                        match pack
-                            .categories
-                            .all_categories
-                            .entry(category.full_id.clone())
-                        {
-                            Entry::Occupied(mut existing) => {
-                                existing.get_mut().merge(category);
-                            }
-                            Entry::Vacant(vacant) => {
-                                vacant.insert(category);
-                            }
-                        }
-                    }
-                    "poi" => {
-                        let Some(PartialItem::Poi(poi)) = parse_stack.pop() else {
-                            anyhow::bail!("Inconsistent internal state");
-                        };
-
-                        pack.pois.push(poi);
-                    }
-                    "trail" => {
-                        let Some(PartialItem::Trail(trail)) = parse_stack.pop() else {
-                            anyhow::bail!("Inconsistent internal state");
-                        };
-
-                        pack.trails.push(trail);
-                    }
-                    _ => anyhow::bail!("Unexpected </{name}>"),
-                }
-            }
-            XmlEvent::EndElement { name } => {
-                anyhow::bail!("Unexpected </{name}>")
-            }
-            XmlEvent::StartDocument { .. } => {}
-            XmlEvent::EndDocument => {
-                if !parse_stack.is_empty() {
-                    anyhow::bail!("Unexpected end of document");
-                }
-                break;
-            }
-            XmlEvent::ProcessingInstruction { .. } => {}
-            XmlEvent::CData(_) => {}
-            XmlEvent::Comment(_) => {}
-            XmlEvent::Characters(_) => {}
-            XmlEvent::Whitespace(_) => {}
-        }
-    }
-    Ok(())
-}
-
-pub enum PartialItem {
-    OverlayData,
-    MarkerCategory(Category),
-    PoiGroup,
-    Poi(poi::Poi),
-    Trail(trail::Trail),
-    PoisonElem,
-}
-
-impl PartialItem {
-    fn as_category(&self) -> Option<&Category> {
-        match self {
-            PartialItem::MarkerCategory(category) => Some(category),
-            _ => None,
-        }
-    }
-
-    fn is_poison(&self) -> bool {
-        match self {
-            PartialItem::PoisonElem => true,
-            _ => false,
+impl MarkerAttributesExt for MarkerAttributes {
+    fn parse_schedule(&self) -> anyhow::Result<Option<Cron>> {
+        match &self.schedule {
+            Some(schedule) => schedule.parse()
+                .context("parsing marker schedule")
+                .map(Some),
+            None => Ok(None),
         }
     }
 }
 
-fn valid_elem_start(stack_top: Option<&PartialItem>, name: &xml::name::OwnedName) -> bool {
-    match (name.local_name.to_ascii_lowercase().as_str(), stack_top) {
-        ("overlaydata", None) => true,
-        ("markercategory", Some(PartialItem::OverlayData | PartialItem::MarkerCategory(_))) => true,
-        ("pois", Some(PartialItem::OverlayData)) => true,
-        ("poi", Some(PartialItem::PoiGroup)) => true,
-        ("trail", Some(PartialItem::PoiGroup)) => true,
-        _ => false,
+pub trait PoiExt {
+    fn marker_position(&self) -> Point3<PackSpace>;
+    fn position(&self) -> Point3<PackSpace>;
+    fn offset(&self) -> Vector3<PackSpace>;
+}
+
+impl PoiExt for Poi {
+    fn offset(&self) -> Vector3<PackSpace> {
+        let height_offset = self.height_offset();
+        Vector3::new(0.0, height_offset, 0.0)
+    }
+
+    #[inline]
+    fn marker_position(&self) -> Point3<PackSpace> {
+        Point3::from_raw(self.position.into())
+    }
+
+    fn position(&self) -> Point3<PackSpace> {
+        self.marker_position() + self.offset()
     }
 }
 
-fn valid_elem_end(stack_top: Option<&PartialItem>, name: &xml::name::OwnedName) -> bool {
-    match (name.local_name.to_ascii_lowercase().as_str(), stack_top) {
-        ("overlaydata", Some(PartialItem::OverlayData)) => true,
-        ("markercategory", Some(PartialItem::MarkerCategory(_))) => true,
-        ("pois", Some(PartialItem::PoiGroup)) => true,
-        ("poi", Some(PartialItem::Poi(_))) => true,
-        ("trail", Some(PartialItem::Trail(_))) => true,
-        _ => false,
+#[cfg(todo)]
+pub trait TrailExt {
+}
+
+#[cfg(todo)]
+impl TrailExt for Trail {
+}
+
+pub trait TrailSectionExt {
+    fn bounds(&self) -> Box3<PackSpace>;
+}
+
+impl TrailSectionExt for TrailSection {
+    #[inline]
+    fn bounds(&self) -> Box3<PackSpace> {
+        let min = Point3::from_raw(self.bounds.min.to_raw());
+        let max = Point3::from_raw(self.bounds.max.to_raw());
+        Box3::new(min, max)
     }
 }
