+++ src/space/pack/trail.rs
@@ -1,155 +1,20 @@
 use {
-    super::{
-        attributes::MarkerAttributes, loader::PackLoaderContext, taco_safe_name, taco_xml_to_guid,
-        ActivePack, Pack,
-    },
-    crate::{
-        marker::atomic::MapSpace,
-        space::{
-            dx11::VertexBuffer,
-            resources::{Model, Texture, Vertex},
-        },
+    crate::space::{
+        dx11::VertexBuffer,
+        pack::{ActivePack, TrailSectionExt},
+        resources::{Model, Texture, Vertex},
+        DrawSpace,
     },
     anyhow::Context,
     core::f32,
-    glamour::{point3, vec3, Box3, Point3, Union, Vector3},
-    std::{io::BufReader, sync::Arc},
-    uuid::Uuid,
-    windows::{
-        core::Interface as _,
-        Win32::Graphics::{
-            Direct3D::D3D_PRIMITIVE_TOPOLOGY_TRIANGLESTRIP,
-            Direct3D11::{ID3D11Buffer, ID3D11Device, ID3D11DeviceContext},
-        },
+    glamour::{Box3, Vector3},
+    std::sync::Arc,
+    taimi_pack::Trail,
+    windows::Win32::Graphics::{
+        Direct3D::D3D_PRIMITIVE_TOPOLOGY_TRIANGLESTRIP,
+        Direct3D11::{ID3D11Device, ID3D11DeviceContext},
     },
 };
-
-pub struct Trail {
-    pub category: String,
-    pub guid: Uuid,
-    pub data: TrailData,
-    pub attributes: MarkerAttributes,
-}
-
-impl Trail {
-    pub fn from_xml(
-        pack: &mut Pack,
-        ctx: &mut impl PackLoaderContext,
-        attrs: Vec<xml::attribute::OwnedAttribute>,
-    ) -> anyhow::Result<Trail> {
-        let mut category = String::new();
-        let mut trail_path = None;
-        let mut guid = None;
-        let mut attributes = MarkerAttributes::default();
-
-        for attr in attrs {
-            if attr.name.local_name.eq_ignore_ascii_case("type") {
-                category = taco_safe_name(&attr.value, true);
-            } else if attr.name.local_name.eq_ignore_ascii_case("traildata") {
-                trail_path = Some(attr.value);
-            } else if attr.name.local_name.eq_ignore_ascii_case("guid") {
-                guid = Some(taco_xml_to_guid(&attr.value));
-            } else if !attributes.try_add(pack, &attr) {
-                log::warn!("Unknown Trail attribute '{}'", attr.name.local_name);
-            }
-        }
-
-        if category.is_empty() {
-            anyhow::bail!("No 'type' specified for Trail");
-        }
-
-        let Some(trail_path) = trail_path else {
-            anyhow::bail!("No 'trailData' specified for Trail '{category}'");
-        };
-
-        let data = read_trl_file(BufReader::new(ctx.load_asset(&trail_path)?), &trail_path)?;
-        let guid = guid.unwrap_or_default();
-
-        Ok(Trail {
-            category,
-            guid,
-            data,
-            attributes,
-        })
-    }
-}
-
-pub struct TrailData {
-    pub map_id: i32,
-    pub sections: Vec<TrailSection>,
-}
-
-pub struct TrailSection {
-    pub points: Vec<Point3<MapSpace>>,
-    pub bounds: Box3<MapSpace>,
-}
-
-pub fn read_trl_file(mut reader: impl std::io::Read, name: &str) -> anyhow::Result<TrailData> {
-    let mut buf32 = [0u8; 4];
-    reader
-        .read_exact(&mut buf32)
-        .context("Reading trail version")?;
-    if i32::from_le_bytes(buf32) != 0 {
-        anyhow::bail!("Trl version '0' is the only known valid format version");
-    }
-
-    reader
-        .read_exact(&mut buf32)
-        .context("Reading trail map_id")?;
-    let map_id = i32::from_le_bytes(buf32);
-
-    let mut sections = vec![];
-    let mut current_section = vec![];
-
-    const NEG_BOX: Box3<MapSpace> = glamour::Box3 {
-        min: point3!(f32::INFINITY, f32::INFINITY, f32::INFINITY),
-        max: point3!(f32::NEG_INFINITY, f32::NEG_INFINITY, f32::NEG_INFINITY),
-    };
-
-    let mut bounds = NEG_BOX;
-    let mut read_more = true;
-    while read_more {
-        let point_data = match read_point(&mut reader) {
-            Ok(point_data) => point_data,
-            Err(e) if e.kind() == std::io::ErrorKind::UnexpectedEof => {
-                read_more = false;
-                EMPTY_POINT
-            }
-            Err(e) => return Err(e).context("Reading trail sections"),
-        };
-
-        if point_data == EMPTY_POINT {
-            if !current_section.is_empty() {
-                sections.push(TrailSection {
-                    points: std::mem::take(&mut current_section),
-                    bounds,
-                });
-                bounds = NEG_BOX;
-            } else {
-                log::warn!("Empty trail section in {name}");
-            }
-        } else {
-            let x = f32::from_le_bytes(point_data[0]);
-            let y = f32::from_le_bytes(point_data[1]);
-            let z = f32::from_le_bytes(point_data[2]);
-            let point = point3!(x, y, z);
-            current_section.push(point);
-            bounds = bounds.union(Box3::new(point, point));
-        }
-    }
-
-    Ok(TrailData { map_id, sections })
-}
-
-const EMPTY_POINT: [[u8; 4]; 3] = [[0; 4]; 3];
-
-fn read_point(reader: &mut impl std::io::Read) -> std::io::Result<[[u8; 4]; 3]> {
-    let mut point_data = [[0; 4]; 3];
-    reader.read_exact(&mut point_data[0])?;
-    reader.read_exact(&mut point_data[1])?;
-    reader.read_exact(&mut point_data[2])?;
-    Ok(point_data)
-}
 
 pub struct ActiveTrail {
     pub category_idx: usize,
@@ -157,7 +22,7 @@ pub struct ActiveTrail {
     pub render_bookmark: usize,
 
     // Segment data.
-    pub section_bounds: Vec<Box3<MapSpace>>,
+    pub section_bounds: Vec<Box3<DrawSpace>>,
 
     // World render data.
     pub texture: Arc<Texture>,
@@ -177,19 +42,12 @@ impl ActiveTrail {
         render_bookmark: usize,
         device: &ID3D11Device,
     ) -> anyhow::Result<ActiveTrail> {
-        let texture_handle = trail
-            .attributes
-            .texture
-            .as_ref()
+        let texture_handle = trail.texture_name()
             .ok_or_else(|| anyhow::anyhow!("TODO: Add a fallback texture for trails"))?;
         let texture_handle = loader.register_texture(texture_handle);
         let texture = loader
             .get_or_load_texture(texture_handle, device)
             .context("Loading trail texture")?;
-
-        let attrs = &trail.attributes;
-        let is_wall = attrs.is_wall.unwrap_or(false);
-        let trail_scale = attrs.trail_scale.unwrap_or(1.0);
 
         let mut vertices: Vec<Vertex> = Vec::new();
         let mut section_bookmarks: Vec<u32> = vec![0];
@@ -231,14 +89,14 @@ impl ActiveTrail {
             let mut cur_point = points[0];
             let mut last_offset = Vector3::ZERO;
             let mut flip_over = 1.0f32;
-            let normal_offset = TRAIL_WIDTH * trail_scale;
+            let normal_offset = TRAIL_WIDTH * trail.scale();
             let mut mod_distance = Vector3::ZERO;
 
             let mut distance = 0.0f32;
             for &next_point in points.iter().skip(1) {
                 let path_direction = next_point - cur_point;
                 let offset = path_direction.cross(Vector3::Y);
-                let offset = if is_wall {
+                let offset = if trail.is_wall() {
                     path_direction.cross(offset)
                 } else {
                     offset
@@ -283,7 +141,7 @@ impl ActiveTrail {
             });
 
             section_bookmarks.push(vertices.len() as u32);
-            section_bounds.push(section.bounds);
+            section_bounds.push(section.bounds());
         }
 
         if vertices.is_empty() {
@@ -309,7 +167,7 @@ impl ActiveTrail {
         })
     }
 
-    pub fn update(pack: &mut Pack, trail_idx: usize) {
+    pub fn update(pack: &mut ActivePack, trail_idx: usize) {
         let _ = pack;
         let _ = trail_idx;
     }
